requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-generativeai>=0.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
bcrypt==4.3.0
